final touch

+++ backend/app/controllers/AnimeC.php
@@ -8,12 +8,11 @@ class AnimeC extends Controller
         $this->anime = $this->model("AnimeM");
     }
 
-
-    public function read()
+    public function readCategorise()
     {
 
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            // $data = json_decode(file_get_contents("php://input"));
+            $data = json_decode(file_get_contents("php://input"));
 
             // $headers = apache_request_headers();
 
@@ -22,8 +21,27 @@ class AnimeC extends Controller
             // if (count($headers) == 1) {
             //     echo json_encode(["response" => "redirect"]);
             // } else {
+            if ($d = $this->anime->readCategorise($data))
+                echo json_encode($d, JSON_INVALID_UTF8_IGNORE);
+            // }
+        }
+    }
+
+    public function read()
+    {
+
+        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+            // $data = json_decode(file_get_contents("php://input"));
+
+            $headers = $this->gettokenFromFront();
+            if (isset($this->valid_token($headers)->iss)) {
                 if ($this->anime->read())
                     echo json_encode($this->anime->read());
+            } else {
+                echo json_encode(['token' => 'token is expired']);
+            }
+            // } else {
+
             // }
         }
     }
@@ -32,7 +50,7 @@ class AnimeC extends Controller
 
         // $d = $this->anime->readAll();
 
-        
+
         // print_r(json_encode((object) $d));
 
 
@@ -40,25 +58,25 @@ class AnimeC extends Controller
         //     ["data" => $this->anime->readAll()]
         //  , JSON_UNESCAPED_UNICODE)) ;
 
-        print_r(json_encode( $this->anime->readAll(), JSON_INVALID_UTF8_IGNORE ));
+        print_r(json_encode($this->anime->readAll(), JSON_INVALID_UTF8_IGNORE));
 
 
 
         // if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-            // $headers = apache_request_headers();
+        // $headers = apache_request_headers();
 
-            // $headers = $headers['authorization'] ? explode(" ", $headers['authorization']) : '';
+        // $headers = $headers['authorization'] ? explode(" ", $headers['authorization']) : '';
 
-            // if (count($headers) == 1) {
-            //     echo json_encode(["response" => "redirect"]);
-            // } else {
+        // if (count($headers) == 1) {
+        //     echo json_encode(["response" => "redirect"]);
+        // } else {
 
-                
-                // }
 
-               
-            // }
+        // }
+
+
+        // }
         // }
     }
     public function readA()
@@ -66,7 +84,7 @@ class AnimeC extends Controller
 
         // $d = $this->anime->readAll();
 
-        
+
         // print_r(json_encode((object) $d));
         $data = json_decode(file_get_contents("php://input"));
 
@@ -75,49 +93,49 @@ class AnimeC extends Controller
         //     ["data" => $this->anime->readAll()]
         //  , JSON_UNESCAPED_UNICODE)) ;
 
-        print_r(json_encode( $this->anime->readA($data), JSON_INVALID_UTF8_IGNORE ));
+        print_r(json_encode($this->anime->readA($data), JSON_INVALID_UTF8_IGNORE));
 
 
 
         // if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-            // $headers = apache_request_headers();
+        // $headers = apache_request_headers();
 
-            // $headers = $headers['authorization'] ? explode(" ", $headers['authorization']) : '';
+        // $headers = $headers['authorization'] ? explode(" ", $headers['authorization']) : '';
 
-            // if (count($headers) == 1) {
-            //     echo json_encode(["response" => "redirect"]);
-            // } else {
+        // if (count($headers) == 1) {
+        //     echo json_encode(["response" => "redirect"]);
+        // } else {
 
-                // die(print_r($this->anime->readAll()));
-                // $d = [];
-                // $i = 0;
-                // while($i < count($this->anime->readAll()))
-                // {
-                //     echo json_encode($this->anime->readAll()[0]);
-                //     echo json_encode($this->anime->readAll()[1]);
-                //     echo json_encode($this->anime->readAll()[2]);
-                //     $i++;
-                // }
+        // die(print_r($this->anime->readAll()));
+        // $d = [];
+        // $i = 0;
+        // while($i < count($this->anime->readAll()))
+        // {
+        //     echo json_encode($this->anime->readAll()[0]);
+        //     echo json_encode($this->anime->readAll()[1]);
+        //     echo json_encode($this->anime->readAll()[2]);
+        //     $i++;
+        // }
 
-                // $age = array("Peter"=>"35", "Ben"=>"37", "Joe"=>"43");
-                // print_r($this->anime->readAll());
-                // echo json_encode();
+        // $age = array("Peter"=>"35", "Ben"=>"37", "Joe"=>"43");
+        // print_r($this->anime->readAll());
+        // echo json_encode();
 
-                // die(print_r($this->anime->readAll()));
-                    // echo json_encode();
+        // die(print_r($this->anime->readAll()));
+        // echo json_encode();
 
-                // echo count($this->anime->readAll());
+        // echo count($this->anime->readAll());
 
-                // die(var_dump($d));
-                // if ()
+        // die(var_dump($d));
+        // if ()
 
-                // die(var_dump($d));
-                    
-            // }
+        // die(var_dump($d));
+
+        // }
         // }
     }
-    
+
     public function read_single()
     {
         $data = json_decode(file_get_contents("php://input"));
@@ -139,22 +157,35 @@ class AnimeC extends Controller
     {
 
         $data = json_decode(file_get_contents("php://input"));
+        $headers = $this->gettokenFromFront();
 
-        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            if ($this->anime->insert($data)) {
-                echo json_encode(["Success", "Created Successfully"]);
+        if (isset($this->valid_token($headers)->iss)) {
+            if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+                if ($this->anime->insert($data)) {
+                    echo json_encode(["Success", "Created Successfully"]);
+                }
             }
+        } else {
+            echo json_encode(['token' => 'token is expired']);
         }
     }
 
     public function delete()
     {
         $data = json_decode(file_get_contents("php://input"));
-        if ($_SERVER['REQUEST_METHOD'] == "DELETE") {
-            if ($this->anime->delete($data)) {
-                echo json_encode(['Success' => 'deleted Successfully']);
+
+        $headers = $this->gettokenFromFront();
+
+        if (isset($this->valid_token($headers)->iss)) {
+            if ($_SERVER['REQUEST_METHOD'] == "DELETE") {
+                if ($this->anime->delete($data)) {
+                    echo json_encode(['Success' => 'deleted Successfully']);
+                }
             }
+        } else {
+            echo json_encode(['token' => 'token is expired']);
         }
+        
     }
 
 
@@ -162,11 +193,18 @@ class AnimeC extends Controller
     {
         $data = json_decode(file_get_contents("php://input"));
 
-        if ($_SERVER['REQUEST_METHOD'] == 'PUT') {
-            if ($this->anime->update($data)) {
-                echo json_encode(['Success' => 'data has updated']);
+        $headers = $this->gettokenFromFront();
+
+        if (isset($this->valid_token($headers)->iss)) {
+            if ($_SERVER['REQUEST_METHOD'] == 'PUT') {
+                if ($this->anime->update($data)) {
+                    echo json_encode(['Success' => 'data has updated']);
+                }
             }
+        } else {
+            echo json_encode(['token' => 'token is expired']);
         }
+        
         // } else {
         //     echo json_encode("exist");
         // }
